… README.md
… src/Parlot/Fluent/OneOrMany.cs
… src/Parlot/Fluent/Separated.cs
… src/Parlot/Fluent/Sequence.cs
… src/Parlot/Fluent/SkipWhiteSpace.cs
… src/Parlot/Fluent/StringLiteral.cs
… src/Parlot/Fluent/TextLiteral.cs
… src/Parlot/Fluent/Then.cs
… src/Parlot/Fluent/ZeroOrMany.cs
… src/Parlot/Fluent/ZeroOrOne.cs
… src/Parlot/IsExternalInit.cs
… src/Parlot/Parlot.csproj
… src/Parlot/Rewriting/IRewritable.cs
… src/Parlot/Rewriting/ISeekable.cs
… src/Parlot/Scanner.cs
… test/Parlot.Benchmarks/ExprBench.cs
… test/Parlot.Benchmarks/JsonBench.cs
… test/Parlot.Benchmarks/Parlot.Benchmarks.csproj
… test/Parlot.Benchmarks/ParlotBenchmarks.cs
… test/Parlot.Benchmarks/PidginParsers/PidginJsonParser.cs
… test/Parlot.Benchmarks/Program.cs
… test/Parlot.Benchmarks/RegexBenchmarks.cs
… test/Parlot.Benchmarks/SpracheParsers/SpracheJsonParser.cs
… test/Parlot.Benchmarks/SuperpowerParsers/SuperpowerJsonParser.cs
… test/Parlot.Tests/CompileTests.cs
… test/Parlot.Tests/CursorTests.cs
… test/Parlot.Tests/Json/JsonModel.cs
… test/Parlot.Tests/Json/JsonParser.cs
… test/Parlot.Tests/Models/FakeSeekable.cs
… test/Parlot.Tests/RewriteTests.cs